+++ ask-sdk-model/ask_sdk_model/interfaces/alexa/presentation/aplt/render_document_directive.py
@@ -34,7 +34,7 @@ class RenderDocumentDirective(Directive):
     :type token: (optional) str
     :param target_profile: One of supported profiles in character display. Default value is NONE.
     :type target_profile: (optional) ask_sdk_model.interfaces.alexa.presentation.aplt.target_profile.TargetProfile
-    :param document: The APL document that the devices need to render a presentation.
+    :param document: Depending on the document type, it represents either an entire APLT document or a reference Link to the document. In a Link object, the value of the &#39;src&#39; should follow a URI format defined like &#39;doc://alexa/aplt/documents/&lt;document_id&gt;&#39;. The &#39;document_id&#39; is a reference to the APLT document that the developer stores through APL Authoring Tool.
     :type document: (optional) dict(str, object)
     :param datasources: Data sources to bind to the document when rendering.
     :type datasources: (optional) dict(str, object)
@@ -65,7 +65,7 @@ class RenderDocumentDirective(Directive):
         :type token: (optional) str
         :param target_profile: One of supported profiles in character display. Default value is NONE.
         :type target_profile: (optional) ask_sdk_model.interfaces.alexa.presentation.aplt.target_profile.TargetProfile
-        :param document: The APL document that the devices need to render a presentation.
+        :param document: Depending on the document type, it represents either an entire APLT document or a reference Link to the document. In a Link object, the value of the &#39;src&#39; should follow a URI format defined like &#39;doc://alexa/aplt/documents/&lt;document_id&gt;&#39;. The &#39;document_id&#39; is a reference to the APLT document that the developer stores through APL Authoring Tool.
         :type document: (optional) dict(str, object)
         :param datasources: Data sources to bind to the document when rendering.
         :type datasources: (optional) dict(str, object)