+++ ask-sdk-model/ask_sdk_model/services/directive/directive_service_client.py
@@ -23,6 +23,7 @@ from ask_sdk_model.services.base_service_client import BaseServiceClient
 from ask_sdk_model.services.api_configuration import ApiConfiguration
 from ask_sdk_model.services.service_client_response import ServiceClientResponse
 from ask_sdk_model.services.api_response import ApiResponse
+from ask_sdk_model.services.utils import user_agent_info
 
 
 
@@ -39,13 +40,16 @@ class DirectiveServiceClient(BaseServiceClient):
     :param api_configuration: Instance of ApiConfiguration
     :type api_configuration: ask_sdk_model.services.api_configuration.ApiConfiguration
     """
-    def __init__(self, api_configuration):
-        # type: (ApiConfiguration) -> None
+    def __init__(self, api_configuration, custom_user_agent=None):
+        # type: (ApiConfiguration, str) -> None
         """
         :param api_configuration: Instance of :py:class:`ask_sdk_model.services.api_configuration.ApiConfiguration`
         :type api_configuration: ask_sdk_model.services.api_configuration.ApiConfiguration
+        :param custom_user_agent: Custom User Agent string provided by the developer.
+        :type custom_user_agent: str
         """
         super(DirectiveServiceClient, self).__init__(api_configuration)
+        self.user_agent = user_agent_info(sdk_version="1.0.0", custom_user_agent=custom_user_agent)
 
     def enqueue(self, send_directive_request, **kwargs):
         # type: (SendDirectiveRequest, **Any) -> Union[ApiResponse, Error]
@@ -82,6 +86,7 @@ class DirectiveServiceClient(BaseServiceClient):
         if 'send_directive_request' in params:
             body_params = params['send_directive_request']
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False