+++ ask-sdk-model/ask_sdk_model/services/ups/ups_service_client.py
@@ -23,6 +23,7 @@ from ask_sdk_model.services.base_service_client import BaseServiceClient
 from ask_sdk_model.services.api_configuration import ApiConfiguration
 from ask_sdk_model.services.service_client_response import ServiceClientResponse
 from ask_sdk_model.services.api_response import ApiResponse
+from ask_sdk_model.services.utils import user_agent_info
 
 
 
@@ -42,13 +43,16 @@ class UpsServiceClient(BaseServiceClient):
     :param api_configuration: Instance of ApiConfiguration
     :type api_configuration: ask_sdk_model.services.api_configuration.ApiConfiguration
     """
-    def __init__(self, api_configuration):
-        # type: (ApiConfiguration) -> None
+    def __init__(self, api_configuration, custom_user_agent=None):
+        # type: (ApiConfiguration, str) -> None
         """
         :param api_configuration: Instance of :py:class:`ask_sdk_model.services.api_configuration.ApiConfiguration`
         :type api_configuration: ask_sdk_model.services.api_configuration.ApiConfiguration
+        :param custom_user_agent: Custom User Agent string provided by the developer.
+        :type custom_user_agent: str
         """
         super(UpsServiceClient, self).__init__(api_configuration)
+        self.user_agent = user_agent_info(sdk_version="1.0.0", custom_user_agent=custom_user_agent)
 
     def get_profile_email(self, **kwargs):
         # type: (**Any) -> Union[ApiResponse, str, Error]
@@ -77,6 +81,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False
@@ -137,6 +142,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False
@@ -197,6 +203,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False
@@ -257,6 +264,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False
@@ -325,6 +333,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False
@@ -393,6 +402,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False
@@ -461,6 +471,7 @@ class UpsServiceClient(BaseServiceClient):
 
         body_params = None
         header_params.append(('Content-type', 'application/json'))
+        header_params.append(('User-Agent', self.user_agent))
 
         # Response Type
         full_response = False